Upload .NET project after level 8

+++ DotNet2025_8430_6589/DalFacade/DO/Sale.cs
@@ -10,7 +10,7 @@ namespace DO
         int Id,
         int ProductId, 
         int RequiredQuantity,
-        int PriceAfterDiscount,
+        double PriceAfterDiscount,
         bool IsForClubMemberOnly,
         DateTime StartDate,
         DateTime EndDate
